Mainview refactor (#17)
* refactoring subviews in own modules

* fmt & clippy

* update package version in toml

* optimise release build
