+++ src/view/mod.rs
@@ -1,2 +1,4 @@
+pub mod edview;
 pub mod mainview;
+pub mod svgview;
 pub mod tabview;
