+++ src/view/svgview.rs
@@ -1,0 +1,11 @@
+use floem::prelude::*;
+
+pub fn svgpreview_container(i_preview_sig: &RwSignal<Vec<u8>>) -> Scroll {
+    let preview_sig = *i_preview_sig;
+    dyn_container(
+        move || preview_sig.get(),
+        move |pv| img(move || pv.to_vec()),
+    )
+    .scroll()
+    .style(|s| s.flex_col().size_full())
+}
