+++ src/view/tabview.rs
@@ -69,7 +69,7 @@ pub fn tabbar_container(
     i_rawstr: &RwSignal<String>,
     i_pngpreview: &RwSignal<Vec<u8>>,
     i_preview_id: ViewId,
-) -> impl IntoView {
+) -> Container {
     let render = render_button(i_rawstr, i_pngpreview, i_preview_id);
     let clear = clear_button(i_doc);
     let save = save_button(*i_rawstr);
